g2llvm: locate src/GhidraToXML.java relative to Ghidra-to-LLVM directory
Prior to this commit, invoking g2llvm.py from a directory outside of
Ghidra-to-LLVM would result in a failure to locate src/GhidraToXML.java.

	ERROR REPORT SCRIPT ERROR: ./src/GhidraToXML.java : Script not found: ./src/GhidraToXML.java (HeadlessAnalyzer)

With this commit, the directory of Ghidra-to-LLVM is determined by
taking the realpath of sys.path[0], which either holds the parent
directory of the g2llvm.py script, or an empty string if invoked
from the interpreter. In both cases, realpath will resolve the full
path to the Ghidra-to-LLVM directory, as realpath on an empty string
uses the current working directory to resolve the full path.